+++ backend/app/db/models/resume.py
@@ -1,7 +1,7 @@
 """简历模型。"""
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, String, Text
+from sqlalchemy import Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base, TimestampMixin
@@ -10,7 +10,7 @@ from app.db.base import Base, TimestampMixin
 class Resume(Base, TimestampMixin):
     __tablename__ = "resumes"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     original_filename: Mapped[str] = mapped_column(String(255), nullable=False)
     storage_path: Mapped[str | None] = mapped_column(String(512), nullable=True)
     resume_text: Mapped[str | None] = mapped_column(Text, nullable=True)
